output: WIP - add verdict as an event type
Besides having the alert's verdict as a part of the alert event, offer
the ability to have it as an independent event type.

Task #6210

TODO:
- update documentation
- improve LogVerdictCondition

Ticket's comment:
When it is enabled, it should default to logging for packets that:
1. have alerts
2. had a new pass action trigger (so the first pass if a flow pass was set)
3. had a new drop action trigger (so the first drop if a flow drop was set)
4. a bypass is triggered

Optionally, it should log for each packet that is dropped or "passed".